requirements.txt
google-genai>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
